+++ string.h
@@ -5,5 +5,5 @@
 
 char *strcpy(char *dest, const char *src);
 void *memset(void *s, int c, u64 sz);
-int strlen(char *s);
+size_t strlen(char *s);
 #endif
